Add `style_lip` as argument to `gridfinityInit`
When used as a library, `grifinity-rebuilt-utility.scad` accepts all the
bin properties to use as arguments to `gridfinityInit`, `cut`, or other
methods, with the sole exception of `style_lip` which must be defined in
the calling file.

This change adds `style_lip` as an argument to `gridfinityInit` to
ensure `gridfinity-rebuilt-utility.scad` can be used as a library.