Add flat-top style baseplates
The current design of the baseplates with a sharp edge at the top of the
base plates is not ideal for printability. Slicers try to approximate
this sharp edge with a thin perimeter line that is usually disjoint from
the perimeters at the grid intersections.  The end result is quite
uneven.

Introduce a new "top style" configuration option which defaults to
"sharp" - the current behavior.  An alternate "flattened" setting
removes two top two layers of the baseplate, leaving a flat surface with
a width of roughly two perimeter lines on top.  This is sliced much more
consistently and the printed result also looks a lot better.  The
functional impact of not having a sharp edge on top is neglible.